users grid

+++ src/users/ViewUser.js
@@ -73,14 +73,12 @@ export default function ViewUser() {
             </div>
             <Link className="btn btn-primary my-3 mx-2" to={"/"}><FontAwesomeIcon icon={faHouse} /></Link>
             <Link className='btn btn-success mx-2' to={`/edituser/${user.id}`}><FontAwesomeIcon icon={faPenToSquare} /></Link>
-            {/* <Link className='btn btn-danger mx-2' onClick={()=>deleteUser(user.id)}><FontAwesomeIcon icon={faTrash} /></Link> */}
-
             <button className='btn btn-danger mx-2' onClick={() => {
-                                    const confirmBox = window.confirm( "Do you really want to delete this USER - " + user.username +" ?")
-                                    if (confirmBox === true){
-                                        deleteUser(user.id,user.username)
-                                        }}}><FontAwesomeIcon icon={faTrash} />
-                                </button>
+                const confirmBox = window.confirm( "Do you really want to delete this USER - " + user.username +" ?")
+                    if (confirmBox === true){
+                        deleteUser(user.id,user.username)
+                    }}}><FontAwesomeIcon icon={faTrash} />
+            </button>
 
             </div>
         </div>
